refactor(format): extract analysis jsonld renderer
